@@ -7,7 +7,14 @@ import { AppComponent } from './app.component';
 import { SharedModule } from './shared/shared.module';
 import { UsuariosModule } from './usuarios/usuarios.module';
 
-import {HttpClientModule} from '@angular/common/http';
+import { HttpClientModule } from '@angular/common/http';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { StoreModule } from '@ngrx/store';
+
+import { appReducers } from './store/app.reducers';
+import { environment } from '../environments/environment.prod';
+import { EffectsModule } from '@ngrx/effects';
+import { EffectsArray } from './store/effects/index';
 
 @NgModule({
   declarations: [
@@ -18,7 +25,14 @@ import {HttpClientModule} from '@angular/common/http';
     AppRoutingModule,
     HttpClientModule,
     SharedModule, 
-    UsuariosModule
+    UsuariosModule, 
+    StoreModule.forRoot( appReducers ), 
+    EffectsModule.forRoot( EffectsArray ),
+    StoreDevtoolsModule.instrument({
+      maxAge: 25, 
+      logOnly: environment.production
+    }),
+    
   ],
   providers: [],
   bootstrap: [AppComponent]
