@@ -11,7 +11,7 @@ const routes: Routes = [
 
 @NgModule({
   imports: [
-    RouterModule.forRoot( routes )
+    RouterModule.forRoot(routes, { relativeLinkResolution: 'legacy' })
   ], 
   exports:[
     RouterModule,
