@@ -12,12 +12,18 @@ export class UsuarioService {
   constructor( private http: HttpClient) {
   }
 
-  getUser(){
+  getUsers(){
+    return this.http.get(`${ this.url }/users/?pera_page=6&delay=3`)
+      .pipe(
+        map( resp => resp['data'] )
+      );
+  }
 
-    return this.http.get(`${ this.url }/users/?pera_page=6`)
-    .pipe(
-      map( resp => resp['data'] )
-    );
+  getUserById( id: string ){
+    return this.http.get(`${ this.url }/users/${ id }`)
+      .pipe(
+        map( resp => resp['data'] )
+      );
   }
 
 }
