@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { UsuarioService } from '../../services/usuario.service';
 import { Usuario } from '../../models/usuario.model';
+import { Store } from '@ngrx/store';
+import { AppState } from '../../store/app.reducers';
+import { cargarUsuarios } from '../../store/actions/usuarios.actions';
 
 @Component({
   selector: 'app-lista',
@@ -9,18 +12,29 @@ import { Usuario } from '../../models/usuario.model';
 })
 export class ListaComponent implements OnInit {
 
+  loading: boolean = false;
+  error: any;
   usuarios: Usuario[] = [];
 
-  constructor( public usuarioService: UsuarioService) { }
+  constructor( 
+    //public usuarioService: UsuarioService
+    private store: Store<AppState>
+  ) { }
 
   ngOnInit(): void {
+    this.store.select('usuarios').subscribe( ({users, loading, error}) => {
+     this.usuarios = users;
+     this.loading = loading;
+     this.error = error;
+    })
+    this.store.dispatch( cargarUsuarios() );
 
-    this.usuarioService.getUser()
+  /*  this.usuarioService.getUser()
       .subscribe( users => {
         console.log( users );
         this.usuarios = users;
       })
-
+  */
   }
 
 }
